Add intersection, difference, and symmetricDifference overlay operations
Extracts shared overlay algorithm infrastructure into Overlay.swift (edge
extraction, R-tree intersection finding, edge splitting, ring building,
polygon assembly) and refactors Union to use it.

New operations:
- intersection(with:gridSize:) — geometric intersection of two polygons
- difference(with:gridSize:) — polygon subtraction (A − B)
- symmetricDifference(with:gridSize:) — geometric XOR

All operations work in EPSG:3857 with automatic reprojection, support
gridSize for coordinate snapping, and handle MultiPolygon inputs.
Boundary test uses makeBoundaryTest helper which compares the two sides
of each edge with a supplied "is inside result" closure.

Tests: 7 intersection, 9 difference, 8 symmetricDifference covering
overlap, containment, non-overlap, identity, holes, MultiPolygon,
gridSize, and antimeridian crossing (10 antimeridian tests total).
